+++ nucleus-support-v4/src/main/java/nucleus/view/NucleusFragmentActivity.java
@@ -71,8 +71,9 @@ public abstract class NucleusFragmentActivity<P extends Presenter> extends Fragm
     }
 
     @Override
-    protected void onPause() {
-        super.onPause();
-        presenterDelegate.onPause(isFinishing());
+    protected void onDestroy() {
+        super.onDestroy();
+        presenterDelegate.onDropView();
+        presenterDelegate.onDestroy(isFinishing());
     }
 }