+++ nucleus-support-v4/src/main/java/nucleus/view/NucleusSupportFragment.java
@@ -69,8 +69,13 @@ public abstract class NucleusSupportFragment<P extends Presenter> extends Fragme
     }
 
     @Override
-    public void onPause() {
-        super.onPause();
-        presenterDelegate.onPause(getActivity().isFinishing());
+    public void onDestroyView() {
+        super.onDestroyView();
+        presenterDelegate.onDropView();
+    }
+
+    @Override
+    public void onDestroy() {
+        presenterDelegate.onDestroy(getActivity().isFinishing());
     }
 }