+++ nucleus-support-v7/src/main/java/nucleus/view/NucleusActionBarActivity.java
@@ -71,8 +71,9 @@ public abstract class NucleusActionBarActivity<P extends Presenter> extends Acti
     }
 
     @Override
-    protected void onPause() {
-        super.onPause();
-        presenterDelegate.onPause(isFinishing());
+    protected void onDestroy() {
+        super.onDestroy();
+        presenterDelegate.onDropView();
+        presenterDelegate.onDestroy(isFinishing());
     }
 }