+++ nucleus/src/main/java/nucleus/view/NucleusFragment.java
@@ -69,8 +69,13 @@ public abstract class NucleusFragment<P extends Presenter> extends Fragment impl
     }
 
     @Override
-    public void onPause() {
-        super.onPause();
-        presenterDelegate.onPause(getActivity().isFinishing());
+    public void onDestroyView() {
+        super.onDestroyView();
+        presenterDelegate.onDropView();
+    }
+
+    @Override
+    public void onDestroy() {
+        presenterDelegate.onDestroy(getActivity().isFinishing());
     }
 }