+++ nucleus/src/test/java/nucleus/view/NucleusActivityTest.java
@@ -104,11 +104,14 @@ public class NucleusActivityTest {
         tested.onCreate(null);
         tested.onResume();
         verify(mockDelegate, times(1)).onResume(tested);
-        tested.onPause();
-        verify(mockDelegate, times(1)).onPause(false);
+        tested.onDestroy();
+        verify(mockDelegate, times(1)).onDropView();
+        verify(mockDelegate, times(1)).onDestroy(false);
         tested.onSaveInstanceState(BundleMock.mock());
         verify(mockDelegate, times(1)).onSaveInstanceState();
         tested.onDestroy();
+        verify(mockDelegate, times(2)).onDropView();
+        verify(mockDelegate, times(2)).onDestroy(false);
         verifyNoMoreInteractions(mockPresenter, mockDelegate, mockFactory);
     }
 
@@ -131,8 +134,8 @@ public class NucleusActivityTest {
     public void testDestroy() throws Exception {
         tested.onCreate(null);
         setUpIsFinishing(true);
-        tested.onPause();
         tested.onDestroy();
-        verify(mockDelegate, times(1)).onPause(true);
+        verify(mockDelegate, times(1)).onDropView();
+        verify(mockDelegate, times(1)).onDestroy(true);
     }
 }